main/requirements.txt
asgiref==3.3.4
bson==0.5.10
certifi==2021.5.30
cffi==1.14.5
cryptography==3.4.7
Django==3.2.4
django-cors-headers==3.7.0
django-rest-knox==4.1.0
djangorestframework==3.12.4
gunicorn==20.1.0
joblib==1.0.1
mysqlclient==2.0.3
numpy==1.21.0
pandas==1.2.5
pycparser==2.20
PyMySQL==1.0.2
python-dateutil==2.8.1
pytz==2021.1
scikit-learn==0.24.2
scipy==1.7.0
six==1.16.0
sklearn==0.0
sqlparse==0.4.1
threadpoolctl==2.2.0
touch==2020.12.3
ujson==4.0.2
values==2020.12.3
waitress==2.0.0
wincertstore==0.2
